fix: resolve CI test failures
- GraphQL schema: Long -> Int (no built-in Long scalar in GraphQL)
- ConversionResult/MutationController: Long -> Integer for fileSize
- GraphQLResolverTest: use real JobTrackingService, expect AWAITING_UPLOAD
- MemoryHealthIndicatorTest: account for hysteresis (3 checks before DOWN)
- ContainerRuntimeHealthIndicatorTest: expect UP for string returns (DOWN only on exception)

+++ src/main/java/com/xtopdf/xtopdf/controllers/graphql/ConversionMutationController.java
@@ -26,7 +26,7 @@ public class ConversionMutationController {
    * REST endpoint with file upload. This serves as a job submission/tracking entry point.</p>
    */
   @MutationMapping
-  public ConversionResult convertFile(@Argument String fileName, @Argument Long fileSize) {
+  public ConversionResult convertFile(@Argument String fileName, @Argument Integer fileSize) {
     String outputFileName = fileName.replaceAll("\\.[^.]+$", ".pdf");
     ConversionJob job = jobTrackingService.submit(fileName, outputFileName, null);
