+++ src/main/java/com/xtopdf/xtopdf/controllers/graphql/ConversionResult.java
@@ -7,7 +7,7 @@ public record ConversionResult(
         String id,
         String fileName,
         String status,
-        Long fileSize,
+        Integer fileSize,
         String submittedAt,
         String completedAt,
         String errorMessage) {}
